requirements.txt
psycopg2-binary
pandas
openpyxl
tkinter